[Fix] vertical Int -> Double로 변경

+++ Sinzak/Scene/Detail/Model/ProductsDetail.swift
@@ -34,6 +34,6 @@ struct ProductsDetail {
     let myPost: Bool
     let complete: Bool
     let width: Double
-    let vertical: Int
+    let vertical: Double
     let height: Double
 }
